requirements.txt
Flask==2.0.1
flask-cors==3.0.10
flask-swagger-ui==3.36.0
python-jose==3.3.0
mysql-connector-python==8.0.26
Werkzeug==2.0.1